ide_assists::assist_context

Struct AssistContext

pub(crate) struct AssistContext<'a> {
    pub(crate) config: &'a AssistConfig,
    pub(crate) sema: Semantics<'a, RootDatabase>,
    frange: FileRange,
    trimmed_range: TextRange,
    source_file: SourceFile,
    token_at_offset: TokenAtOffset<SyntaxToken>,
    covering_element: SyntaxElement,
}

AssistContext allows to apply an assist or check if it could be applied.

Assists use a somewhat over-engineered approach, given the current needs. The assists workflow consists of two phases. In the first phase, a user asks for the list of available assists. In the second phase, the user picks a particular assist and it gets applied.

There are two peculiarities here:

first, we ideally avoid computing more things then necessary to answer “is assist applicable” in the first phase.

second, when we are applying assist, we don’t have a guarantee that there weren’t any changes between the point when user asked for assists and when they applied a particular assist. So, when applying assist, we need to do all the checks from scratch.

To avoid repeating the same code twice for both “check” and “apply” functions, we use an approach reminiscent of that of Django’s function based views dealing with forms. Each assist receives a runtime parameter, resolve. It first check if an edit is applicable (potentially computing info required to compute the actual edit). If it is applicable, and resolve is true, it then computes the actual edit.

So, to implement the original assists workflow, we can first apply each edit with resolve = false, and then applying the selected edit again, with resolve = true this time.

Note, however, that we don’t actually use such two-phase logic at the moment, because the LSP API is pretty awkward in this place, and it’s much easier to just compute the edit eagerly :-)

Fields
config: &'a AssistConfigsema: Semantics<'a, RootDatabase>frange: FileRangetrimmed_range: TextRangesource_file: SourceFiletoken_at_offset: TokenAtOffset<SyntaxToken>covering_element: SyntaxElement
Implementations

impl<'a> AssistContext<'a>

pub(crate) fn new( sema: Semantics<'a, RootDatabase>, config: &'a AssistConfig, frange: FileRange, ) -> AssistContext<'a>

pub(crate) fn db(&self) -> &'a RootDatabase

pub(crate) fn offset(&self) -> TextSize

pub(crate) fn vfs_file_id(&self) -> FileId

pub(crate) fn file_id(&self) -> EditionedFileId

pub(crate) fn edition(&self) -> Edition

pub(crate) fn has_empty_selection(&self) -> bool

pub(crate) fn selection_trimmed(&self) -> TextRange

Returns the selected range trimmed for whitespace tokens, that is the range will be snapped to the nearest enclosed token.

pub(crate) fn source_file(&self) -> &SourceFile

pub(crate) fn token_at_offset(&self) -> TokenAtOffset<SyntaxToken>

pub(crate) fn find_token_syntax_at_offset( &self, kind: SyntaxKind, ) -> Option<SyntaxToken>

pub(crate) fn find_token_at_offset<T: AstToken>(&self) -> Option<T>

pub(crate) fn find_node_at_offset<N: AstNode>(&self) -> Option<N>

pub(crate) fn find_node_at_trimmed_offset<N: AstNode>(&self) -> Option<N>

pub(crate) fn find_node_at_range<N: AstNode>(&self) -> Option<N>

pub(crate) fn find_node_at_offset_with_descend<N: AstNode>(&self) -> Option<N>

pub(crate) fn covering_element(&self) -> SyntaxElement

Returns the element covered by the selection range, this excludes trailing whitespace in the selection.