ide_assists::handlers

Module convert_tuple_return_type_to_struct

Functions

add_tuple_struct_def ğŸ”’

augment_references_with_imports ğŸ”’

convert_tuple_return_type_to_struct ğŸ”’

node_to_pats ğŸ”’

replace_body_return_values ğŸ”’

Replaces each returned tuple in body with the constructor of the tuple struct named struct_name.

replace_usages ğŸ”’

Replaces tuple usages with the corresponding tuple struct pattern.

tail_cb_impl ğŸ”’