ide_assists::handlers

Module generate_documentation_template

Functions

all_parent_mods_public ğŸ”’

Checks that all parent modules of the function are public / exported

arguments_from_params ğŸ”’

Helper function to build the comma-separated list of arguments of the function

build_path ğŸ”’

Helper function to build the path of the module in the which is the node

can_panic ğŸ”’

None if function without a body; some bool to guess if function can panic

count_parameters ğŸ”’

Helper function to count the parameters including self

crate_name ğŸ”’

Returns the name of the current crate

documentation_from_lines ğŸ”’

Helper function to transform lines of documentation into a Rust code documentation

errors_builder ğŸ”’

Builds an optional # Errors section

function_call ğŸ”’

Helper function to build a function call. None if expected self_name was not provided

generate_doc_example ğŸ”’

generate_documentation_template ğŸ”’

introduction_builder ğŸ”’

is_a_ref_mut_param ğŸ”’

Helper function to determine if a parameter is &mut

is_in_trait_def ğŸ”’

Helper function to determine if the function definition is in a trait definition

is_in_trait_impl ğŸ”’

Helper function to determine if the function is in a trait implementation

is_public ğŸ”’

Checks if the function is public / exported

is_ref_mut_self ğŸ”’

Returns None if no self at all, Some(true) if there is &mut self else Some(false)

make_example_for_fn ğŸ”’

panics_builder ğŸ”’

Builds an optional # Panics section

ref_mut_params ğŸ”’

Helper function to build the list of &mut parameters

return_type ğŸ”’

Helper function to get the return type of a function

returns_a_value ğŸ”’

Helper function to determine if the function returns some data

safety_builder ğŸ”’

Builds an optional # Safety section

self_name ğŸ”’

Helper function to get the name that should be given to self arguments

self_partial_type ğŸ”’

Helper function to get the name of the type of self without generic arguments

self_type ğŸ”’

Helper function to get the name of the type of self

self_type_without_lifetimes ğŸ”’

Output the real name of Self like MyType<T>, without the lifetimes.

string_vec_from ğŸ”’

Helper function to transform an array of borrowed strings to an owned Vec<String>