ide_assists::handlers::generate_function

Function filter_unnecessary_bounds

fn filter_unnecessary_bounds(
    generic_params: &mut Vec<ParamBoundWithParams>,
    where_preds: &mut Vec<WherePredWithParams>,
    necessary_params: FxHashSet<GenericParam>,
)

Filters out irrelevant bounds from generic_params and where_preds.

Say we have a trait bound Struct<T>: Trait<U>. Given necessary_params, when is it relevant and when not? Some observations:

When necessary_params contains T, it’s likely that we want this bound, but now we have an extra param to consider: U.

On the other hand, when necessary_params contains U (but not T), then it’s unlikely that we want this bound because it doesn’t really constrain U.

(FIXME?: The latter clause might be overstating. We may want to include the bound if the self type does not include generic params at all - like Option<i32>: From<U>)

Can we make this a bit more formal? Let’s define “dependency” between generic parameters and trait bounds:

A generic parameter T depends on a trait bound if T appears in the self type (i.e. left part) of the bound.

A trait bound depends on a generic parameter T if T appears in the bound.

Using the notion, what we want is all the bounds that params in necessary_params transitively depend on!

Now it’s not hard to solve: we build a dependency graph and compute all reachable nodes from nodes that represent params in necessary_params by usual and boring DFS.

The time complexity is O(|generic_params| + |where_preds| + |necessary_params|).