ide_assists::handlers::generate_function

Function make_return_type

fn make_return_type(
    ctx: &AssistContext<'_>,
    expr: &Expr,
    target_module: Module,
    necessary_generic_params: &mut FxHashSet<GenericParam>,
) -> (Option<RetType>, bool)

Makes an optional return type along with whether the return type should be focused by the cursor. If we cannot infer what the return type should be, we create a placeholder type.

The rule for whether we focus a return type or not (and thus focus the function body), is rather simple:

If we could not infer what the return type should be, focus it (so the user can fill-in the correct return type).

If we could infer the return type, don’t focus it (and thus focus the function body) so the user can change the todo! function body.