ide_assists::handlers::inline_type_alias

Function create_replacement

fn create_replacement(
    lifetime_map: &LifetimeMap,
    const_and_type_map: &ConstAndTypeMap,
    concrete_type: &Type,
) -> SyntaxNode

This doesn’t attempt to ensure specified generics are compatible with those required by the type alias, other than lifetimes which must either all be specified or all omitted. It will replace TypeArgs with ConstArgs and vice versa if they’re in the wrong position. It supports partially specified generics.

Map the provided instance’s generic args to the type alias’s generic params:

type A<'a, const N: usize, T = u64> = &'a [T; N];
      ^ alias generic params
let a: A<100>;
        ^ instance generic args

generic['a] = '_ due to omission
generic[N] = 100 due to the instance arg
generic[T] = u64 due to the default param

Copy the concrete type and substitute in each found mapping:

&'_ [u64; 100]

Remove wildcard lifetimes entirely:

&[u64; 100]