ide_assists::handlers

Module replace_if_let_with_match

Functions

and_bin_expr_left ğŸ”’

binds_name ğŸ”’

is_empty_expr ğŸ”’

is_sad_pat ğŸ”’

let_and_guard ğŸ”’

make_else_arm ğŸ”’

pick_pattern_and_expr_order ğŸ”’

Pick the pattern for the if let condition and return the expressions for the then body and else body in that order.

replace_if_let_with_match ğŸ”’

replace_match_with_if_let ğŸ”’