Function ide_assists::utils::test_related_attribute

pub fn test_related_attribute(fn_def: &Fn) -> Option<Attr>

This is a method with a heuristics to support test methods annotated with custom test annotations, such as #[test_case(...)], #[tokio::test] and similar. Also a regular #[test] annotation is supported.

It may produce false positives, for example, #[wasm_bindgen_test] requires a different command to run the test, but it’s better than not to have the runnables for the tests at all.