ide_assists::utils

Module gen_trait_fn_body

This module contains functions to generate default trait impl function bodies where possible.

Functions

gen_clone_impl ğŸ”’

Generate a Clone impl based on the fields and members of the target type.

gen_debug_impl ğŸ”’

Generate a Debug impl based on the fields and members of the target type.

gen_default_impl ğŸ”’

Generate a Debug impl based on the fields and members of the target type.

gen_hash_impl ğŸ”’

Generate a Hash impl based on the fields and members of the target type.

gen_partial_eq ğŸ”’

Generate a PartialEq impl based on the fields and members of the target type.

gen_partial_ord ğŸ”’

gen_trait_fn_body ğŸ”’

Generate custom trait bodies without default implementation where possible.

make_discriminant ğŸ”’