ide_assists::utils

Module ref_field_expr

This module contains a helper for converting a field access expression into a path expression. This is used when destructuring a tuple or struct.

It determines whether to deref the new expression and/or wrap it in parentheses, based on the parent of the existing expression.

Structs

RefData 🔒

Indicates whether to deref an expression or wrap it in parens

Functions

determine_ref_and_parens 🔒

Decides whether the new path expression needs to be dereferenced and/or wrapped in parens. Returns the relevant parent expression to replace and the RefData.