ide_assists
Module utils

Assorted functions shared by several assists.

Enums

DefaultMethods

IgnoreAssocItems

Functions

add_trait_assoc_items_to_impl

Given original_items retrieved from the trait definition (usually by filter_assoc_items()), clones each item for update and applies path transformation to it, then inserts into impl_. Returns the modified impl_ and the first associated item that got inserted.

extract_trivial_expression

filter_assoc_items

has_test_related_attribute

is_body_const

test_related_attribute_syn

This is a method with a heuristics to support test methods annotated with custom test annotations, such as #[test_case(...)], #[tokio::test] and similar. Also a regular #[test] annotation is supported.